feat: validate parcels are connected
