+++ packages/@dcl/ecs/src/runtime/helpers/index.ts
@@ -1,1 +1,2 @@
+export * from './coordinates'
 export * from './tree'
