+++ test/sdk-commands/utils/scene-validation.spec.ts
@@ -1,6 +1,6 @@
+import { areConnected } from '../../../packages/@dcl/ecs/src/runtime/helpers/coordinates'
 import { initComponents } from '../../../packages/@dcl/sdk-commands/src/components'
 import * as v from '../../../packages/@dcl/sdk-commands/src/logic/scene-validations'
-import { areConnected } from '../../../packages/@dcl/sdk-commands/src/logic/coordinates'
 import * as projectFiles from '../../../packages/@dcl/sdk-commands/src/logic/project-files'
 import { Stats } from 'fs'
