+++ test/snapshots/production-bundles/billboard.ts.crdt
@@ -77,4 +77,4 @@ CALL onUpdate(0.1)
   OPCODES ~= 8k
   MALLOC_COUNT = 0
   ALIVE_OBJS_DELTA ~= 0.00k
-  MEMORY_USAGE_COUNT ~= 881.34k bytes
+  MEMORY_USAGE_COUNT ~= 881.38k bytes
