+++ test/snapshots/production-bundles/cubes.ts.crdt
@@ -1652,4 +1652,4 @@ CALL onUpdate(0.1)
   OPCODES ~= 636k
   MALLOC_COUNT = 0
   ALIVE_OBJS_DELTA ~= 0.00k
-  MEMORY_USAGE_COUNT ~= 1019.54k bytes
+  MEMORY_USAGE_COUNT ~= 1019.58k bytes
