+++ test/snapshots/production-bundles/ui.ts.crdt
@@ -65,4 +65,4 @@ CALL onUpdate(0.1)
   OPCODES ~= 61k
   MALLOC_COUNT = 0
   ALIVE_OBJS_DELTA ~= 0.00k
-  MEMORY_USAGE_COUNT ~= 1511.33k bytes
+  MEMORY_USAGE_COUNT ~= 1511.37k bytes
